crispr pipeline - on all modalities
Line 112 in metadata.py was edited since the merge was giving an error that object and float64 columns could not merged. Hence we added a column name Metadata_JCP2022 for the well metadata to be merged with orf and crsipr. The second error was found in line 79 of stats.py. This was fixed by commenting out that line since John confirmed that it was not used downstream