Upadted crispr pipeline_small batch
changed the metadata.py 'get_well_metadata' function according to John's suggestions and ran the updated crispr pipeline (with cell count) for a smaller batch. The cell count csv of all the modalities were concatenated and there were no duplicates to drop and the csv was named as all_cell_counts. all_cell_counts csv was used for the crispr pipeline.